// src/types/user.ts
export interface User {
  id: string
  orgName: string
  userName: string
  email: string
  phoneNumber: string
  dateJoined: string
  status: 'Active' | 'Inactive' | 'Pending' | 'Blacklisted'
  profile: {
    firstName: string
    lastName: string
    phoneNumber: string
    avatar: string
    gender: string
    bvn: string
    address: string
    currency: string
  }
  guarantor: {
    firstName: string
    lastName: string
    phoneNumber: string
    gender: string
    address: string
  }
  accountBalance: string
  accountNumber: string
  socials: {
    facebook: string
    instagram: string
    twitter: string
  }
  education: {
    level: string
    employmentStatus: string
    sector: string
    duration: string
    officeEmail: string
    monthlyIncome: string[]
    loanRepayment: string
  }
}

export interface UserStats {
  totalUsers: number
  activeUsers: number
  usersWithLoans: number
  usersWithSavings: number
}